+++ HISTORY.rst
@@ -33,6 +33,8 @@ latest (unreleased)
 * QoQa Package types are now "delivery.carrier"
 * Forbid usage of attribute values with more than 25 chars. Historic values
   might still be longer but are not allowed to be used.
+* QoQa users are no longer imported as companies, now Odoo 9 allows an
+  individual to have addresses
 
 **Bugfixes**